+++ src/EntityGraphQL/Compiler/EntityQuery/EntityQueryNodeVisitor.cs
@@ -115,7 +115,7 @@ namespace EntityGraphQL.Compiler.EntityQuery
                 return enumOrConstantValue;
             }
             var gqlField = schemaProvider.GetActualField(schemaType.Name, field, requestContext);
-            var exp = gqlField.GetExpression(currentContext, null);
+            var exp = gqlField.GetExpression(gqlField.Resolve, currentContext, null, null, false);
             return exp;
         }