+++ src/EntityGraphQL/Compiler/GqlNodes/BaseGraphQLField.cs
@@ -26,6 +26,10 @@ namespace EntityGraphQL.Compiler
         public Expression NextFieldContext { get; set; }
         public IGraphQLNode ParentNode { get; set; }
         public ParameterExpression RootParameter { get; set; }
+        /// <summary>
+        /// Arguments from inline in the query
+        /// </summary>
+        protected readonly Dictionary<string, Expression> arguments;
 
         /// <summary>
         /// Name of the field
@@ -33,12 +37,13 @@ namespace EntityGraphQL.Compiler
         /// <value></value>
         public string Name { get; protected set; }
 
-        public BaseGraphQLField(string name, Expression nextFieldContext, ParameterExpression rootParameter, IGraphQLNode parentNode)
+        public BaseGraphQLField(string name, Expression nextFieldContext, ParameterExpression rootParameter, IGraphQLNode parentNode, Dictionary<string, Expression> arguments)
         {
             Name = name;
             NextFieldContext = nextFieldContext;
             RootParameter = rootParameter;
             ParentNode = parentNode;
+            this.arguments = arguments;
         }
 
         /// <summary>
@@ -48,7 +53,7 @@ namespace EntityGraphQL.Compiler
         protected Dictionary<ParameterExpression, object> constantParameters = new();
         public List<BaseGraphQLField> QueryFields { get; } = new();
         internal Dictionary<ParameterExpression, object> ConstantParameters { get => constantParameters; }
-        public List<Type> Services { get; } = new List<Type>();
+        public List<Type> Services { get; set; } = new List<Type>();
         /// <summary>
         /// Field is a complex expression (using a method or function) that returns a single object (not IEnumerable)
         /// We wrap this is a function that does a null check and avoid duplicate calls on the method/service
@@ -66,12 +71,13 @@ namespace EntityGraphQL.Compiler
         /// </summary>
         /// <param name="serviceProvider">Service provider to resolve services </param>
         /// <param name="fragments">Fragments in the query document</param>
+        /// <param name="parentArguments">Inline arguments from the parent fields. May be needed when a Field Extension rebuilds the schema shape</param>
         /// <param name="withoutServiceFields">If true the expression builds without fields that require services</param>
         /// <param name="replacementNextFieldContext">A replacement context from a selection without service fields</param>
         /// <param name="isRoot">If this field is a Query root field</param>
         /// <param name="contextChanged">If true the context has changed. This means we are compiling/executing against the result ofa pre-selection without service fields</param>
         /// <returns></returns>
-        public abstract Expression GetNodeExpression(IServiceProvider serviceProvider, List<GraphQLFragmentStatement> fragments, ParameterExpression schemaContext, bool withoutServiceFields, Expression replacementNextFieldContext = null, bool isRoot = false, bool contextChanged = false);
+        public abstract Expression GetNodeExpression(IServiceProvider serviceProvider, List<GraphQLFragmentStatement> fragments, Dictionary<string, Expression> parentArguments, ParameterExpression schemaContext, bool withoutServiceFields, Expression replacementNextFieldContext = null, bool isRoot = false, bool contextChanged = false);
 
         public abstract IEnumerable<BaseGraphQLField> Expand(List<GraphQLFragmentStatement> fragments, bool withoutServiceFields);
 
@@ -104,13 +110,13 @@ namespace EntityGraphQL.Compiler
             return (baseExpression, listTypeParam);
         }
 
-        protected (Expression baseExpression, Dictionary<string, CompiledField> selectionExpressions, ParameterExpression selectContextParam) ProcessExtensionsSelection(GraphQLFieldType fieldType, Expression baseExpression, Dictionary<string, CompiledField> selectionExpressions, ParameterExpression selectContextParam, ParameterReplacer parameterReplacer)
+        protected (Expression baseExpression, Dictionary<string, CompiledField> selectionExpressions, ParameterExpression selectContextParam) ProcessExtensionsSelection(GraphQLFieldType fieldType, Expression baseExpression, Dictionary<string, CompiledField> selectionExpressions, ParameterExpression selectContextParam, bool servicesPass, ParameterReplacer parameterReplacer)
         {
             if (fieldExtensions != null)
             {
                 foreach (var extension in fieldExtensions)
                 {
-                    (baseExpression, selectionExpressions, selectContextParam) = extension.ProcessExpressionSelection(fieldType, baseExpression, selectionExpressions, selectContextParam, parameterReplacer);
+                    (baseExpression, selectionExpressions, selectContextParam) = extension.ProcessExpressionSelection(fieldType, baseExpression, selectionExpressions, selectContextParam, servicesPass, parameterReplacer);
                 }
             }
             return (baseExpression, selectionExpressions, selectContextParam);
@@ -129,9 +135,12 @@ namespace EntityGraphQL.Compiler
 
         internal void AddConstantParameters(IReadOnlyDictionary<ParameterExpression, object> constantParameters)
         {
+            if (constantParameters == null)
+                return;
             foreach (var item in constantParameters)
             {
-                this.constantParameters.Add(item.Key, item.Value);
+                if (!this.constantParameters.ContainsKey(item.Key))
+                    this.constantParameters.Add(item.Key, item.Value);
             }
         }
     }