+++ src/EntityGraphQL/Compiler/GqlNodes/GraphQLMutationField.cs
@@ -9,21 +9,19 @@ namespace EntityGraphQL.Compiler
     public class GraphQLMutationField : BaseGraphQLQueryField
     {
         private readonly MutationType mutationType = null;
-        private readonly Dictionary<string, Expression> args = null;
         public BaseGraphQLQueryField ResultSelection { get; set; }
 
         public GraphQLMutationField(string name, MutationType mutationType, Dictionary<string, Expression> args, Expression nextFieldContext, ParameterExpression rootParameter, IGraphQLNode parentNode)
-            : base(name, nextFieldContext, rootParameter, parentNode)
+            : base(name, nextFieldContext, rootParameter, parentNode, args)
         {
             this.mutationType = mutationType;
-            this.args = args;
         }
 
         public async Task<object> ExecuteMutationAsync<TContext>(TContext context, GraphQLValidator validator, IServiceProvider serviceProvider, Func<string, string> fieldNamer)
         {
             try
             {
-                return await mutationType.CallAsync(context, args, validator, serviceProvider, fieldNamer);
+                return await mutationType.CallAsync(context, arguments, validator, serviceProvider, fieldNamer);
             }
             catch (EntityQuerySchemaException e)
             {
@@ -31,9 +29,9 @@ namespace EntityGraphQL.Compiler
             }
         }
 
-        public override Expression GetNodeExpression(IServiceProvider serviceProvider, List<GraphQLFragmentStatement> fragments, ParameterExpression schemaContext, bool withoutServiceFields, Expression replacementNextFieldContext = null, bool isRoot = false, bool contextChanged = false)
+        public override Expression GetNodeExpression(IServiceProvider serviceProvider, List<GraphQLFragmentStatement> fragments, Dictionary<string, Expression> parentArguments, ParameterExpression schemaContext, bool withoutServiceFields, Expression replacementNextFieldContext = null, bool isRoot = false, bool contextChanged = false)
         {
-            return ResultSelection.GetNodeExpression(serviceProvider, fragments, schemaContext, withoutServiceFields, replacementNextFieldContext);
+            return ResultSelection.GetNodeExpression(serviceProvider, fragments, parentArguments, schemaContext, withoutServiceFields, replacementNextFieldContext);
         }
     }
 }