+++ src/EntityGraphQL/Compiler/GqlNodes/GraphQLObjectProjectionField.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using EntityGraphQL.Compiler.Util;
+using EntityGraphQL.Extensions;
+using EntityGraphQL.Schema;
 using EntityGraphQL.Schema.FieldExtensions;
 
 namespace EntityGraphQL.Compiler
@@ -23,7 +25,7 @@ namespace EntityGraphQL.Compiler
     /// </summary>
     public class GraphQLObjectProjectionField : BaseGraphQLQueryField
     {
-        private readonly ExpressionExtractor extractor;
+        private readonly IField field;
 
         /// <summary>
         /// Create a new GraphQLQueryNode. Represents both fields in the query as well as the root level fields on the Query type
@@ -33,11 +35,12 @@ namespace EntityGraphQL.Compiler
         /// <param name="nextFieldContext">The next context expression for ObjectProjection is also our field expression e..g person.manager</param>
         /// <param name="rootParameter">The root parameter</param>
         /// <param name="parentNode"></param>
-        public GraphQLObjectProjectionField(List<IFieldExtension> extensions, string name, Expression nextFieldContext, ParameterExpression rootParameter, IGraphQLNode parentNode)
-            : base(name, nextFieldContext, rootParameter, parentNode)
+        public GraphQLObjectProjectionField(IField field, List<IFieldExtension> extensions, string name, Expression nextFieldContext, ParameterExpression rootParameter, IGraphQLNode parentNode, Dictionary<string, Expression> arguments)
+            : base(name, nextFieldContext, rootParameter, parentNode, arguments)
         {
             this.fieldExtensions = extensions;
-            extractor = new ExpressionExtractor();
+            this.field = field;
+            this.AddServices(field.Services);
         }
 
         /// <summary>
@@ -46,12 +49,11 @@ namespace EntityGraphQL.Compiler
         /// If there is a object selection (new {} in a Select() or not) we will build the NodeExpression on
         /// Execute() so we can look up any query fragment selections
         /// </summary>
-        public override Expression GetNodeExpression(IServiceProvider serviceProvider, List<GraphQLFragmentStatement> fragments, ParameterExpression schemaContext, bool withoutServiceFields, Expression replacementNextFieldContext = null, bool isRoot = false, bool contextChanged = false)
+        public override Expression GetNodeExpression(IServiceProvider serviceProvider, List<GraphQLFragmentStatement> fragments, Dictionary<string, Expression> parentArguments, ParameterExpression schemaContext, bool withoutServiceFields, Expression replacementNextFieldContext = null, bool isRoot = false, bool contextChanged = false)
         {
-            bool needsServiceWrap = !withoutServiceFields && HasAnyServices(fragments);
-
             var nextFieldContext = NextFieldContext;
-            if (contextChanged && Name != "__typename")
+
+            if (contextChanged && Name != "__typename" && replacementNextFieldContext != null)
             {
                 var possibleField = replacementNextFieldContext.Type.GetField(Name);
                 if (possibleField != null)
@@ -59,44 +61,37 @@ namespace EntityGraphQL.Compiler
                 else
                     nextFieldContext = isRoot ? replacementNextFieldContext : replacer.ReplaceByType(nextFieldContext, ParentNode.NextFieldContext.Type, replacementNextFieldContext);
             }
+            var nextFieldContextExp = field.GetExpression(nextFieldContext, replacementNextFieldContext ?? ParentNode.NextFieldContext, schemaContext, parentArguments.MergeNew(arguments), contextChanged);
+            nextFieldContext = nextFieldContextExp.Expression;
+            AddServices(nextFieldContextExp.Services);
+            AddConstantParameters(nextFieldContextExp.ConstantParameters);
+            bool needsServiceWrap = !withoutServiceFields && HasAnyServices(fragments);
+
 
             (nextFieldContext, _) = ProcessExtensionsPreSelection(GraphQLFieldType.ObjectProjection, nextFieldContext, null, replacer);
+
+            var selectionFields = GetSelectionFields(serviceProvider, fragments, withoutServiceFields, nextFieldContext, schemaContext, contextChanged);
+            if (selectionFields == null || !selectionFields.Any())
+                return null;
 
             if (needsServiceWrap ||
                 ((nextFieldContext.NodeType == ExpressionType.MemberInit || nextFieldContext.NodeType == ExpressionType.New) && isRoot))
             {
-                var updatedExpression = WrapWithNullCheck(serviceProvider, fragments, withoutServiceFields, nextFieldContext, schemaContext, contextChanged);
-                nextFieldContext = updatedExpression;
+                nextFieldContext = WrapWithNullCheck(selectionFields, serviceProvider, nextFieldContext, schemaContext, contextChanged);
             }
             else
             {
-                var selectionFields = GetSelectionFields(serviceProvider, fragments, withoutServiceFields, nextFieldContext, schemaContext, contextChanged);
-
-                if (selectionFields == null || !selectionFields.Any())
-                    return null;
-
-                (nextFieldContext, selectionFields, _) = ProcessExtensionsSelection(GraphQLFieldType.ObjectProjection, nextFieldContext, selectionFields, null, replacer);
+                (nextFieldContext, selectionFields, _) = ProcessExtensionsSelection(GraphQLFieldType.ObjectProjection, nextFieldContext, selectionFields, null, contextChanged, replacer);
                 // build a new {...} - returning a single object {}
                 var newExp = ExpressionUtil.CreateNewExpression(selectionFields.ExpressionOnly(), out Type anonType);
                 if (nextFieldContext.NodeType != ExpressionType.MemberInit && nextFieldContext.NodeType != ExpressionType.New)
                 {
-                    if (selectionFields.Any() && !withoutServiceFields)
-                    {
-                        // make a null check from this new expression
-                        newExp = Expression.Condition(Expression.MakeBinary(ExpressionType.Equal, nextFieldContext, Expression.Constant(null)), Expression.Constant(null, anonType), newExp, anonType);
-                        nextFieldContext = newExp;
-                    }
-                    else
-                    {
-                        nextFieldContext = Expression.Condition(Expression.MakeBinary(ExpressionType.Equal, nextFieldContext, Expression.Constant(null)), Expression.Constant(null, anonType), newExp, anonType);
-                    }
+                    // make a null check from this new expression
+                    nextFieldContext = Expression.Condition(Expression.MakeBinary(ExpressionType.Equal, nextFieldContext, Expression.Constant(null)), Expression.Constant(null, anonType), newExp, anonType);
                 }
                 else
                 {
-                    if (selectionFields.Any() && !withoutServiceFields)
-                        nextFieldContext = newExp;
-                    else
-                        nextFieldContext = newExp;
+                    nextFieldContext = newExp;
                 }
             }
 
@@ -119,14 +114,8 @@ namespace EntityGraphQL.Compiler
         /// <param name="replacementNextFieldContext"></param>
         /// <param name="schemaContext"></param>
         /// <returns></returns>
-        private Expression WrapWithNullCheck(IServiceProvider serviceProvider, List<GraphQLFragmentStatement> fragments, bool withoutServiceFields, Expression nextFieldContext, ParameterExpression schemaContext, bool contextChanged)
+        private Expression WrapWithNullCheck(Dictionary<string, CompiledField> selectionFields, IServiceProvider serviceProvider, Expression nextFieldContext, ParameterExpression schemaContext, bool contextChanged)
         {
-            // don't replace context is needsServiceWrap as the selection fields happen internally to the wrap call on the correct context
-            var selectionFields = GetSelectionFields(serviceProvider, fragments, withoutServiceFields, nextFieldContext, schemaContext, contextChanged);
-
-            if (selectionFields == null || !selectionFields.Any())
-                return null;
-
             // selectionFields is set up but we need to wrap
             // we wrap here as we have access to the values and services etc
             var fieldParamValues = new List<object>(ConstantParameters.Values);
@@ -134,17 +123,13 @@ namespace EntityGraphQL.Compiler
 
             var updatedExpression = Services.Any() ? GraphQLHelper.InjectServices(serviceProvider, Services, fieldParamValues, nextFieldContext, fieldParams, replacer) : nextFieldContext;
             // replace with null_wrap
+            // this is the parameter used in the null wrap. We pass it to the wrap function which has the value to match
             var nullWrapParam = Expression.Parameter(updatedExpression.Type, "nullwrap");
-            // This is the var the we use in the select - the result of the service at runtime
-            var selectionParams = new List<ParameterExpression>();
-            selectionParams.AddRange(fieldParams);
             var selectionParamValues = new List<object>();
             selectionParamValues.AddRange(fieldParamValues);
 
             if (contextChanged)
             {
-                nullWrapParam = Expression.Parameter(updatedExpression.Type, "nullwrap");
-
                 foreach (var item in selectionFields)
                 {
                     if (item.Value.Field.Services.Any())
@@ -162,11 +147,11 @@ namespace EntityGraphQL.Compiler
                 }
             }
 
-            (updatedExpression, selectionFields, _) = ProcessExtensionsSelection(GraphQLFieldType.ObjectProjection, updatedExpression, selectionFields, null, replacer);
+            (updatedExpression, selectionFields, _) = ProcessExtensionsSelection(GraphQLFieldType.ObjectProjection, updatedExpression, selectionFields, null, contextChanged, replacer);
             // we need to make sure the wrap can resolve any services in the select
             var selectionExpressions = selectionFields.ToDictionary(f => f.Key, f => GraphQLHelper.InjectServices(serviceProvider, f.Value.Field.Services, fieldParamValues, f.Value.Expression, fieldParams, replacer));
 
-            updatedExpression = ExpressionUtil.WrapFieldForNullCheck(updatedExpression, selectionParams, selectionExpressions, selectionParamValues, nullWrapParam, schemaContext);
+            updatedExpression = ExpressionUtil.WrapFieldForNullCheck(updatedExpression, fieldParams, selectionExpressions, selectionParamValues, nullWrapParam, schemaContext);
             return updatedExpression;
         }
 
@@ -175,11 +160,16 @@ namespace EntityGraphQL.Compiler
             // fieldExpression might be a service method call and the arguments might have fields we need to select out
             if (withoutServiceFields && NextFieldContext.NodeType == ExpressionType.Call)
             {
-                IDictionary<string, Expression> fieldsRequiredForServices = extractor.Extract(NextFieldContext, RootParameter);
+                var extractor = new ExpressionExtractor();
+                IDictionary<string, Expression> fieldsRequiredForServices = extractor.Extract(NextFieldContext, ParentNode.NextFieldContext, true);
                 if (fieldsRequiredForServices != null)
                 {
                     var fields = fieldsRequiredForServices
-                        .Select(i => new GraphQLScalarField(null, i.Key, i.Value, RootParameter, ParentNode))
+                        .Select(i => new GraphQLScalarField((Field)field, null, i.Key, i.Value, RootParameter, ParentNode, arguments)
+                        {
+                            // do not push services into the fields extracted from a service field
+                            Services = new List<Type>()
+                        })
                         .ToList();
 
                     if (fields.Any())