+++ src/EntityGraphQL/Compiler/GqlNodes/GraphQLScalarField.cs
@@ -3,23 +3,26 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using EntityGraphQL.Compiler.Util;
+using EntityGraphQL.Extensions;
+using EntityGraphQL.Schema;
 using EntityGraphQL.Schema.FieldExtensions;
 
 namespace EntityGraphQL.Compiler
 {
     public class GraphQLScalarField : BaseGraphQLField
     {
-        private readonly ExpressionExtractor extractor;
         private readonly ParameterReplacer replacer;
         private List<GraphQLScalarField> extractedFields;
+        private readonly Field field;
 
-        public GraphQLScalarField(IEnumerable<IFieldExtension> fieldExtensions, string name, Expression nextFieldContext, ParameterExpression rootParameter, IGraphQLNode parentNode)
-            : base(name, nextFieldContext, rootParameter, parentNode)
+        public GraphQLScalarField(Field field, IEnumerable<IFieldExtension> fieldExtensions, string name, Expression nextFieldContext, ParameterExpression rootParameter, IGraphQLNode parentNode, Dictionary<string, Expression> arguments)
+            : base(name, nextFieldContext, rootParameter, parentNode, arguments)
         {
             this.fieldExtensions = fieldExtensions?.ToList();
             Name = name;
-            extractor = new ExpressionExtractor();
             replacer = new ParameterReplacer();
+            this.field = field;
+            this.AddServices(field.Services);
         }
 
         public override bool HasAnyServices(IEnumerable<GraphQLFragmentStatement> fragments)
@@ -43,17 +46,27 @@ namespace EntityGraphQL.Compiler
             if (extractedFields != null)
                 return extractedFields;
 
-            extractedFields = extractor.Extract(NextFieldContext, RootParameter)?.Select(i => new GraphQLScalarField(null, i.Key, i.Value, RootParameter, ParentNode)).ToList();
+            var extractor = new ExpressionExtractor();
+            extractedFields = extractor.Extract(NextFieldContext, ParentNode.NextFieldContext, true)?.Select(i => new GraphQLScalarField(field, null, i.Key, i.Value, RootParameter, ParentNode, arguments)
+            {
+                // do not carry the services over
+                Services = new List<Type>()
+            }).ToList();
             return extractedFields;
         }
 
-        public override Expression GetNodeExpression(IServiceProvider serviceProvider, List<GraphQLFragmentStatement> fragments, ParameterExpression schemaContext, bool withoutServiceFields, Expression replacementNextFieldContext = null, bool isRoot = false, bool contextChanged = false)
+        public override Expression GetNodeExpression(IServiceProvider serviceProvider, List<GraphQLFragmentStatement> fragments, Dictionary<string, Expression> parentArguments, ParameterExpression schemaContext, bool withoutServiceFields, Expression replacementNextFieldContext = null, bool isRoot = false, bool contextChanged = false)
         {
             if (withoutServiceFields && Services.Any())
                 return null;
 
-            var newExpression = NextFieldContext;
-            if (contextChanged && Name != "__typename")
+            var result = field.GetExpression(NextFieldContext, replacementNextFieldContext ?? ParentNode.NextFieldContext, schemaContext, parentArguments.MergeNew(arguments), contextChanged);
+            AddConstantParameters(result.ConstantParameters);
+            AddServices(result.Services);
+
+            var newExpression = result.Expression;
+
+            if (contextChanged && Name != "__typename" && replacementNextFieldContext != null)
             {
                 var selectedField = replacementNextFieldContext.Type.GetField(Name);
                 if (!Services.Any() && selectedField != null)