+++ src/EntityGraphQL/Schema/FieldExtensions/Filter/FilterExtension.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Linq;
 using System.Linq.Expressions;
-using EntityGraphQL.Compiler;
 using EntityGraphQL.Compiler.Util;
 using EntityGraphQL.Extensions;
 
@@ -31,11 +30,17 @@ namespace EntityGraphQL.Schema.FieldExtensions
             isQueryable = typeof(IQueryable).IsAssignableFrom(field.Resolve.Type);
         }
 
-        public override Expression GetExpression(Field field, Expression expression, ParameterExpression argExpression, dynamic arguments, Expression context, ParameterReplacer parameterReplacer)
+        public override Expression GetExpression(Field field, Expression expression, ParameterExpression argExpression, dynamic arguments, Expression context, bool servicesPass, ParameterReplacer parameterReplacer)
         {
+            // data is already filtered
+            if (servicesPass)
+                return expression;
+
             // we have current context update Items field
-            if (arguments.Filter != null && arguments.Filter.HasValue)
-                expression = Expression.Call(isQueryable ? typeof(Queryable) : typeof(Enumerable), "Where", new Type[] { listType }, expression, arguments.Filter.Query);
+            if (arguments != null && arguments?.Filter != null && arguments?.Filter.HasValue)
+            {
+                expression = Expression.Call(isQueryable ? typeof(Queryable) : typeof(Enumerable), "Where", new Type[] { listType }, expression, arguments!.Filter.Query);
+            }
 
             return expression;
         }